Sessão 17: Palestrantes e Redes Sociais - FrontEnd

+++ Back/src/ProEventos.API/helpers/Util.cs
@@ -34,10 +34,13 @@ namespace ProEventos.API.helpers
         [NonAction]
         public void DeleteImage(string imageName, string destino)
         {
-            var imagePath = Path.Combine(_hostEnv.ContentRootPath, @$"Resources/{destino}", imageName);
-            if (System.IO.File.Exists(imagePath))
+            if (!string.IsNullOrEmpty(imageName))
             {
-                System.IO.File.Delete(imagePath);
+                var imagePath = Path.Combine(_hostEnv.ContentRootPath, @$"Resources/{destino}", imageName);
+                if (System.IO.File.Exists(imagePath))
+                {
+                    System.IO.File.Delete(imagePath);
+                }
             }
         }
     }
